+++ CHANGELOG.md
@@ -5,6 +5,11 @@ Tutte le modifiche significative a questo progetto sono documentate in questo fi
 Il formato segue le linee guida di [Keep a Changelog](https://keepachangelog.com/it-IT/1.0.0/)
 e adotta il [versionamento semantico](https://semver.org/lang/it/).
 
+## [1.3.0] - 2025-05-02
+### Added
+- Aggiunta l'opzione `-o <percorso>` alla riga di comando per permettere
+  all’utente di specificare la directory di destinazione dei file di output.
+
 ## [1.2.1] - 2025-05-01
 ### Fixed
 - Modificata la classe `CNFConverter` per evitare la semplificazione automatica